+++ tests/db_test.py
@@ -1,6 +1,13 @@
 
-from wn import _db
+import sqlite3
+
+import pytest
+
+import wn
 
 
+@pytest.mark.usefixtures('mini_db')
 def test_schema_compatibility():
-    assert _db.is_schema_compatible(create=True)
+    conn = sqlite3.connect(wn.config.database_path)
+    schema_hash = wn._db.schema_hash(conn)
+    assert schema_hash in wn._db.COMPATIBLE_SCHEMA_HASHES